==> 1-data-trimming.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.preprocessing import OneHotEncoder

## Load the data

dia_data0=pd.read_csv('dataset_diabetes/diabetic_data.csv')
ids_map=pd.read_csv('dataset_diabetes/IDs_mapping.csv')

## Look at some of the data

Seems it is all categorical?

dia_data0.head()

dia_data0['diag_1']

## Check for number of unique values for each potential feature

potential_features=dia_data0.columns
for n in potential_features:
    num_uniq=len(dia_data0[n].unique())
    print('Number of unique '+ n + ': '+str(num_uniq))

## Delete uninformative features

The 101,766 patient encounters all have the same value for 2 features:

'examide' and 'citoglipton'

These two features will not be useful so we can delete them. 

We will also delete features with many many unique but uninformative values (diag_1, diag_2, and diag_3) for simplicity.

We will also delete any features including '?' or missing values (race, weight, payer_code, medical_specialty).

Later we will also delete from training data the patient id and encounter numbers and the Y value we want to predict ('encounter_id', 'patient_nbr','readmitted').

dia_data=dia_data0.drop(['examide','citoglipton','weight','diag_1','diag_2','diag_3','race','payer_code','medical_specialty'], axis=1)
dia_data.shape

potential_features=dia_data.columns
for n in potential_features:
    num_uniq=len(dia_data[n].unique())
    print('Number of unique '+ n + ': '+str(num_uniq))
    print(dia_data[n].unique())

## Save trimmed data

Will use OneHotEncoder from sklearn... later 

https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OneHotEncoder.html

Potential features (X) will be all remaining features except identifiers 'encounter_id' and 'patient_nbr' and variable we want to predict, readmission so these are removed below.

Y=dia_data['readmitted'].to_numpy()
encounter_id=dia_data['encounter_id'].to_numpy()
patient_nbr=dia_data['patient_nbr'].to_numpy()
X_df=dia_data.drop(['encounter_id', 'patient_nbr','readmitted'], axis=1)

X_array=X_df.to_numpy()

np.save('dataset_diabetes/X_array.npy',X_array)
np.save('dataset_diabetes/Y.npy',Y)
np.save('dataset_diabetes/encounter_id.npy',encounter_id)
np.save('dataset_diabetes/patient_nbr.npy',patient_nbr)
