==> 5-Neural-Net.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.preprocessing import OneHotEncoder
from sklearn.decomposition import TruncatedSVD
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix

## Load the data

X_SVD=np.load('dataset_diabetes/SVD_selected_25_X.npy',allow_pickle=True)
X_tree=np.load('dataset_diabetes/tree_selected_25_X.npy',allow_pickle=True)
readmission=np.load('dataset_diabetes/Y.npy',allow_pickle=True)
encounter_id=np.load('dataset_diabetes/encounter_id.npy',allow_pickle=True)
patient_nbr=np.load('dataset_diabetes/patient_nbr.npy',allow_pickle=True)

my_list = np.where(readmission == 'NO', 0, readmission)
my_list2 = np.where(my_list == '>30', 0, my_list)
Y0 = np.where(my_list2 == '<30', 1, my_list2)
Y=list(Y0)

## Split into 2/3 training and 1/3 testing data sets

X0_train, X0_test, Y0_train, Y0_test = train_test_split(X_SVD, Y, test_size=0.33, random_state=1)
X1_train, X1_test, Y1_train, Y1_test = train_test_split(X_tree, Y, test_size=0.33, random_state=1)

print(np.shape(X0_train))
print(np.shape(X1_train))

## Train Neural Network

Used 10 hidden layers with 100 nodes each for deep architecture. Set max iterations to be high and tolerance to be low to give each network ample training time to converge. Report final loss values for each network.

n_nodes = 10;
n_layers = 100;
nnet_SVD = MLPClassifier(hidden_layer_sizes=tuple(np.full(n_nodes,n_layers)),random_state=1, max_iter=10000,tol=1e-8).fit(X0_train,Y0_train)
nnet_tree = MLPClassifier(hidden_layer_sizes=tuple(np.full(n_nodes,n_layers)),random_state=1, max_iter=10000,tol=1e-8).fit(X1_train,Y1_train)

print('SVD Loss: '+str(nnet_SVD.loss_))
print('Tree Loss: '+str(nnet_tree.loss_))

## Confusion Matrices

Use to analyze how well classification was done for readmitted cases, since these are arguably more important to get correct.

Yhat_SVD_train = nnet_SVD.predict(X0_train)
Yhat_tree_train = nnet_tree.predict(X1_train)
Yhat_SVD_test = nnet_SVD.predict(X0_test)
Yhat_tree_test = nnet_tree.predict(X1_test)

CM_SVD_Train = confusion_matrix(Y0_train,Yhat_SVD_train);
CM_Tree_Train = confusion_matrix(Y1_train,Yhat_tree_train);
CM_SVD_Test = confusion_matrix(Y0_test,Yhat_SVD_test);
CM_Tree_Test = confusion_matrix(Y1_test,Yhat_tree_test);

print('SVD Train Conf. Matrix: '+'\n'+str(CM_SVD_Train)+'\n')
print('Tree Train Conf. Matrix: '+'\n'+str(CM_Tree_Train)+'\n')

print('SVD Test Conf. Matrix: '+'\n'+str(CM_SVD_Test)+'\n')
print('Tree Test Conf. Matrix: '+'\n'+str(CM_Tree_Test))

SVD_train_total = np.sum(CM_SVD_Train,axis=1)[1];
SVD_train_correct = CM_SVD_Train[1][1];

Tree_train_total = np.sum(CM_Tree_Train,axis=1)[1];
Tree_train_correct = CM_Tree_Train[1][1];

print('SVD Training Data Set:')
print('Total Readmitted Patients: '+str(SVD_train_total))
print('Total Correctly Predicted: '+str(SVD_train_correct)+'\n')
print('Tree Training Data Set:')
print('Total Readmitted Patients: '+str(Tree_train_total))
print('Total Correctly Predicted: '+str(Tree_train_correct))

SVD_test_total = np.sum(CM_SVD_Test,axis=1)[1];
SVD_test_correct = CM_SVD_Test[1][1];

Tree_test_total = np.sum(CM_Tree_Test,axis=1)[1];
Tree_test_correct = CM_Tree_Test[1][1];

print('SVD Training Data Set:')
print('Total Readmitted Patients: '+str(SVD_test_total))
print('Total Correctly Predicted: '+str(SVD_test_correct)+'\n')
print('Tree Training Data Set:')
print('Total Readmitted Patients: '+str(Tree_test_total))
print('Total Correctly Predicted: '+str(Tree_test_correct))
